Extract shared interactive element configuration into helper.
