+++ lib/rules/lint-invalid-interactive.js
@@ -2,9 +2,15 @@
 
 var buildPlugin = require('./base');
 var isInteractiveElement = require('../helpers/is-interactive-element');
+var parseConfig = require('../helpers/parse-interactive-element-config');
+var AstNodeInfo = require('../helpers/ast-node-info');
 
 module.exports = function(addonContext) {
   var InvalidInteractive = buildPlugin(addonContext, 'invalid-interactive');
+
+  InvalidInteractive.prototype.parseConfig = function(config) {
+    return parseConfig(this.ruleName, config);
+  };
 
   InvalidInteractive.prototype.visitors = function() {
     this._element = null;
