+++ lib/rules/lint-nested-interactive.js
@@ -18,33 +18,9 @@
 
 var buildPlugin = require('./base');
 var isInteractiveElement = require('../helpers/is-interactive-element');
+var parseConfig = require('../helpers/parse-interactive-element-config');
 
 var ARRAY_DEPRECATION_MESSAGE = 'Specifying an array as the configurate for the `nested-interactive` rule is deprecated and will be removed in future versions.  Please update `.template-lintrc.js` to use the newer object format.';
-
-function configValid(config) {
-  for (var key in config) {
-    var value = config[key];
-
-    switch (key) {
-    case 'ignoredTags':
-    case 'additionalInteractiveTags':
-      if (!Array.isArray(value)) {
-        return false;
-      }
-      break;
-    case 'ignoreTabindex':
-    case 'ignoreUsemapAttribute':
-      if (typeof value !== 'boolean') {
-        return false;
-      }
-      break;
-    default:
-      return false;
-    }
-  }
-
-  return true;
-}
 
 function convertConfigArrayToObject(config) {
   var base = {
@@ -75,40 +51,11 @@ module.exports = function(addonContext) {
   var LogNestedInteractive = buildPlugin(addonContext, 'nested-interactive');
 
   LogNestedInteractive.prototype.parseConfig = function(config) {
-    var configType = typeof config;
-
-    var errorMessage = [
-      'The nested-interactive rule accepts one of the following values.',
-      '  * boolean - `true` to enable / `false` to disable',
-      '  * object - Containing the following values:',
-      '    * `ignoredTags` - An array of element tag names that should be whitelisted. Default to `[]`.',
-      '    * `ignoreTabindex` - When `true` tabindex will be ignored. Defaults to `false`.',
-      '    * `ignoreUsemapAttribute` - When `true` ignores the `usemap` attribute on `img` and `object` elements. Defaults `false`.',
-      '    * `additionalInteractiveTags` - An array of element tag names that should also be considered as interactive. Defaults to `[]`.',
-      'You specified `' + JSON.stringify(config) + '`'
-    ].join('\n');
-
-    switch (configType) {
-    case 'boolean':
-      return config;
-    case 'object':
-      if (configValid(config)) {
-        return config;
-      } else if (Array.isArray(config)) {
-        this.log({
-          message: ARRAY_DEPRECATION_MESSAGE,
-          source: JSON.stringify(config),
-          severity: 1
-        });
-        return convertConfigArrayToObject(config);
-      } else {
-        throw new Error(errorMessage);
-      }
-    case 'undefined':
-      return false;
-    default:
-      throw new Error(errorMessage);
+    if (Array.isArray(config)) {
+      return convertConfigArrayToObject(config);
     }
+
+    return parseConfig(this.ruleName, config);
   };
 
   LogNestedInteractive.prototype.getConfigWhiteList = function() {
